+++ l10n_it_einvoice_in/README.rst
@@ -219,12 +219,13 @@ An Enhancement Proposal may be submitted if your idea gains ground.
 ChangeLog History | Cronologia modifiche
 ----------------------------------------
 
-10.0.1.3.48 (2024-07-16)
+10.0.1.3.48 (2024-07-20)
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
 * [FIX] Wrong invoice total if wrong rounding / Totale fattura errato in caso di errato arrotondamento
 * [IMP] Store and show e-invoice totals / Memorizza e mostra totali e-fattura
-* [QUA] Test coverage 64% (1361: 485+876) [0 TestPoints] - quality rating 39 (target 100)
+* [IMP] Memorizza dati terzo intermediario
+* [QUA] Test coverage 55% (1357: 605+752) [0 TestPoints] - quality rating 34 (target 100)
 
 10.0.1.3.47 (2024-06-14)
 ~~~~~~~~~~~~~~~~~~~~~~~~
@@ -359,7 +360,7 @@ La distribuzione `Zeroincombenze® <https://www.zeroincombenze.it/>`__ è proget
 
 This module is part of l10n-italy project.
 
-Last Update / Ultimo aggiornamento: 2024-07-16
+Last Update / Ultimo aggiornamento: 2024-07-20
 
 .. |Maturity| image:: https://img.shields.io/badge/maturity-Alfa-black.png
     :target: https://odoo-community.org/page/development-status
