+++ l10n_it_einvoice_in/static/description/index.html
@@ -208,12 +208,13 @@ La videta mostra solo i file non registrati ma &#232; possibile modificare il fi
     <div class="oe_row oe_spaced">
         <h3 align="center" style="color:#162832;font-size: 26px;font-weight: 300;">ChangeLog History | Cronologia modifiche</h3>
         <div class="oe_span12">
-<h4>10.0.1.3.48 (2024-07-16)</h4>
+<h4>10.0.1.3.48 (2024-07-20)</h4>
 <p align="justify">
 <ul>
 <li>[FIX] Wrong invoice total if wrong rounding / Totale fattura errato in caso di errato arrotondamento</li>
 <li>[IMP] Store and show e-invoice totals / Memorizza e mostra totali e-fattura</li>
-<li>[QUA] Test coverage 64% (1361: 485+876) [0 TestPoints] - quality rating 39 (target 100)</li>
+<li>[IMP] Memorizza dati terzo intermediario</li>
+<li>[QUA] Test coverage 55% (1357: 605+752) [0 TestPoints] - quality rating 34 (target 100)</li>
 </ul>
 </p><p align="justify">
 <h4>10.0.1.3.47 (2024-06-14)</h4>
@@ -331,6 +332,6 @@ La videta mostra solo i file non registrati ma &#232; possibile modificare il fi
     </div>
     <div class="oe_row oe_spaced">
         <p>This module is part of <a href="https://github.com/zeroincombenze/l10n-italy">l10n-italy</a> project.</p>
-        <p>Published information on | Informazioni pubblicate: 2024-07-16</p>
+        <p>Published information on | Informazioni pubblicate: 2024-07-20</p>
     </div>
 </section>
