+++ l10n_it_einvoice_in/tests/logs/show-log.sh
@@ -1,1 +1,1 @@
-less -R $(readlink -f $(dirname $0))/l10n_it_einvoice_in_20240716.txt
+less -R $(readlink -f $(dirname $0))/l10n_it_einvoice_in_20240720.txt
