+++ l10n_it_einvoice_in/tests/logs/zero10.l10n-italy.l10n_it_einvoice_in.conf
@@ -52,5 +52,5 @@ without_demo = False
 workers = 0
 xmlrpc = True
 xmlrpc_interface = 
-xmlrpc_port = 40258
+xmlrpc_port = 38993
